requirements.txt
Flask==2.2.2
replicate
newspaper3k